+++ CHANGELOG.md
@@ -7,6 +7,12 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Changed
+
+- Dedicated awards management page.
+- Unified login.
+- CSS tweaks for readability.
+
 ## [v21.0.5](https://github.com/lexicalunit/spellbot/releases/tag/v21.0.5) - 2026-06-13
 
 ### Fixed